requirements.txt
discord==1.7.3
python-dotenv==0.19.2
requests==2.26.0
waitress==2.0.0
Flask==2.3.2
colorlog==6.8.2